+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Editing and adding curved lanelet via matplotlib visualization
 - Handling projection strings which contain elements which third-party does not support
 - Support for visualizing scenarios with 3D coordinates on 2D plane
+- cr2lanelet line marking conversion
 
 ## [0.7.2] - 2023-07-29